requirements.txt
streamlit
langchain-openai
langchain
faiss-cpu
python-dotenv